+++ src/exec/job.rs
@@ -17,6 +17,7 @@ pub fn job(
     out_perm: Arc<Mutex<()>>,
     show_filesystem_errors: bool,
     buffer_output: bool,
+    dry_run: bool,
 ) -> ExitCode {
     let mut results: Vec<ExitCode> = Vec::new();
     loop {
@@ -39,7 +40,12 @@ pub fn job(
         // Drop the lock so that other threads can read from the receiver.
         drop(lock);
         // Generate a command, execute it and store its exit code.
-        results.push(cmd.generate_and_execute(&value, Arc::clone(&out_perm), buffer_output))
+        results.push(cmd.generate_and_execute(
+            &value,
+            Arc::clone(&out_perm),
+            buffer_output,
+            dry_run,
+        ))
     }
     // Returns error in case of any error.
     merge_exitcodes(results)
@@ -51,6 +57,7 @@ pub fn batch(
     show_filesystem_errors: bool,
     buffer_output: bool,
     limit: usize,
+    dry_run: bool,
 ) -> ExitCode {
     let paths = rx.iter().filter_map(|value| match value {
         WorkerResult::Entry(path) => Some(path),
@@ -63,14 +70,14 @@ pub fn batch(
     });
     if limit == 0 {
         // no limit
-        return cmd.generate_and_execute_batch(paths, buffer_output);
+        return cmd.generate_and_execute_batch(paths, buffer_output, dry_run);
     }
 
     let mut exit_codes = Vec::new();
     let mut peekable = paths.peekable();
     while peekable.peek().is_some() {
         let limited = peekable.by_ref().take(limit);
-        let exit_code = cmd.generate_and_execute_batch(limited, buffer_output);
+        let exit_code = cmd.generate_and_execute_batch(limited, buffer_output, dry_run);
         exit_codes.push(exit_code);
     }
     merge_exitcodes(exit_codes)